+++ workspaces/entity-feedback/packages/app/CHANGELOG.md
@@ -1,5 +1,12 @@
 # app
 
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies [a9b673e]
+  - @backstage-community/plugin-entity-feedback@0.3.3
+
 ## 0.0.2
 
 ### Patch Changes